pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Third project for hexlet. Flask app."
authors = ["Artemy Ablaev <artemyablaev.gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.2"
gunicorn = "^21.2.0"
flake8 = "^7.0.0"
pytest = "^8.1.1"
pytest-cov = "^5.0.0"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
validators = "^0.25.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
